LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-marker-focus-visible-expected.txt



FAIL CSS Overflow Test: ::scroll-marker supports :focus-visible assert_equals: ::scroll-marker supports :focus-visible expected "rgb(0, 128, 0)" but got ""
